ci: stabilise ruff config + apply ruff format
The CI lint job had 69 ruff errors against a near-empty ruff config.
Inventory of fixes (now all green locally):

- pyproject.toml [tool.ruff]
  - line-length 120 (industry-standard for ML code)
  - target-version py310
  - lint select = ['E', 'F']  (I/isort surfaced 100+ noise; defer)
  - per-file-ignores for legacy patterns:
      F401 on optional-dep try-imports (sklearn, torch shims)
      E402 on framework-after-logger setup files
      E741 on /l/ vars in model files (PatchTST literature uses 'l')
      F841 on test assignment-for-side-effect patterns
      E501 on a few unwrap-able literature strings

- ruff format applied to 60 files (mostly quote-style + trailing-comma
  normalisation — single quotes already the project default; no
  semantic changes)

- Auto-fixed 46 issues via ruff --fix (unused imports + similar)

Next CI matrix run on this branch should be green for lint.

+++ liulian/data/noise.py
@@ -95,9 +95,7 @@ def add_impulse_noise(
         return noisy
     spike_idx = np.random.choice(n_samples, n_spikes, replace=False)
     flat = noisy.ravel()
-    flat[spike_idx] += np.random.choice([-magnitude, magnitude], size=n_spikes).astype(
-        flat.dtype
-    )
+    flat[spike_idx] += np.random.choice([-magnitude, magnitude], size=n_spikes).astype(flat.dtype)
     return noisy
